requirements.txt
gym==0.26.2
imageio==2.22.3
matplotlib==3.6.1
numpy==1.23.3
opencv_python==4.6.0.66
Pillow==9.3.0
pygame==2.1.2
torch==1.12.1
tqdm==4.64.1
